dnsdist: Move HTTP rules to dnsdist-rules.cc
