+++ pdns/dnsdistdist/dnsdist-doh-common.cc
@@ -21,87 +21,9 @@
  */
 #include "base64.hh"
 #include "dnsdist-doh-common.hh"
-#include "dnsdist-rules.hh"
+#include "dnsdist.hh"
 
 #ifdef HAVE_DNS_OVER_HTTPS
-
-HTTPHeaderRule::HTTPHeaderRule(const std::string& header, const std::string& regex) :
-  d_header(toLower(header)), d_regex(regex), d_visual("http[" + header + "] ~ " + regex)
-{
-}
-
-bool HTTPHeaderRule::matches(const DNSQuestion* dq) const
-{
-  if (dq->ids.du) {
-    const auto& headers = dq->ids.du->getHTTPHeaders();
-    for (const auto& header : headers) {
-      if (header.first == d_header) {
-        return d_regex.match(header.second);
-      }
-    }
-    return false;
-  }
-  if (dq->ids.doh3u) {
-    const auto& headers = dq->ids.doh3u->getHTTPHeaders();
-    for (const auto& header : headers) {
-      if (header.first == d_header) {
-        return d_regex.match(header.second);
-      }
-    }
-    return false;
-  }
-  return false;
-}
-
-string HTTPHeaderRule::toString() const
-{
-  return d_visual;
-}
-
-HTTPPathRule::HTTPPathRule(std::string path) :
-  d_path(std::move(path))
-{
-}
-
-bool HTTPPathRule::matches(const DNSQuestion* dq) const
-{
-  if (dq->ids.du) {
-    const auto path = dq->ids.du->getHTTPPath();
-    return d_path == path;
-  }
-  if (dq->ids.doh3u) {
-    return dq->ids.doh3u->getHTTPPath() == d_path;
-  }
-  return false;
-}
-
-string HTTPPathRule::toString() const
-{
-  return "url path == " + d_path;
-}
-
-HTTPPathRegexRule::HTTPPathRegexRule(const std::string& regex) :
-  d_regex(regex), d_visual("http path ~ " + regex)
-{
-}
-
-bool HTTPPathRegexRule::matches(const DNSQuestion* dq) const
-{
-  if (dq->ids.du) {
-    const auto path = dq->ids.du->getHTTPPath();
-    return d_regex.match(path);
-  }
-  if (dq->ids.doh3u) {
-    return d_regex.match(dq->ids.doh3u->getHTTPPath());
-  }
-  return false;
-}
-
-string HTTPPathRegexRule::toString() const
-{
-  return d_visual;
-}
-
 void DOHFrontend::rotateTicketsKey(time_t now)
 {
   return d_tlsContext.rotateTicketsKey(now);
